requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.50.0
django-cors-headers==3.10.1
django-filter==22.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.2
gunicorn==20.1.0
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==5.3.0
zipp==3.8.1
